split sidebar into smaller components
