+++ web/src/shared/components/ContextualHelp/ContextualHelpSidebar/ContextualHelpSidebar.tsx
@@ -1,41 +1,13 @@
-import clsx from 'clsx';
-import { useState } from 'react';
-import { m } from '../../../../paraglide/messages';
-import { Fold } from '../../../defguard-ui/components/Fold/Fold';
-import { Icon, IconKind } from '../../../defguard-ui/components/Icon';
-import { RenderMarkdown } from '../../../defguard-ui/components/RenderMarkdown/RenderMarkdown';
+import { isPresent } from '../../../defguard-ui/utils/isPresent';
 import { useContextualHelp } from '../hooks';
 import type { ContextualHelpKey } from '../types';
+import { ContextualHelpBestPractices } from './ContextualHelpBestPractices';
+import { ContextualHelpDocs } from './ContextualHelpDocs';
+import { ContextualHelpFaqs } from './ContextualHelpFaqs';
 import './ContextualHelpSidebar.scss';
-import { ThemeVariable } from '../../../defguard-ui/types';
-import { isPresent } from '../../../defguard-ui/utils/isPresent';
 
 type Props = {
   pageKey: ContextualHelpKey;
-};
-
-type FaqItemProps = {
-  question: string;
-  answer: string;
-};
-
-const FaqItem = ({ question, answer }: FaqItemProps) => {
-  const [open, setOpen] = useState(false);
-
-  return (
-    <div className="faq-item" onClick={() => setOpen((v) => !v)} role="button">
-      <Icon
-        icon={open ? IconKind.MinusCircle : IconKind.PlusCircle}
-        staticColor={open ? ThemeVariable.FgNeutral : ThemeVariable.FgAction}
-      />
-      <div className="content">
-        <p className={clsx('question', { open })}>{question}</p>
-        <Fold open={open}>
-          <p className="answer">{answer}</p>
-        </Fold>
-      </div>
-    </div>
-  );
 };
 
 export const ContextualHelpSidebar = ({ pageKey }: Props) => {
@@ -50,66 +22,9 @@ export const ContextualHelpSidebar = ({ pageKey }: Props) => {
 
   return (
     <div className="contextual-help">
-      {isPresent(faqs) && (
-        <div className="contextual-help-section">
-          <div className="header">
-            <Icon icon={IconKind.Chat} />
-            <span className="title">{m.cmp_contextual_help_faq()}</span>
-          </div>
-          <div className="faq-card">
-            {faqs.map((faq, i) => (
-              <FaqItem key={i} question={faq.question} answer={faq.answer} />
-            ))}
-          </div>
-        </div>
-      )}
-
-      {isPresent(relatedDocs) && (
-        <div className="contextual-help-section">
-          <div className="header">
-            <Icon icon={IconKind.File} />
-            <span className="title">{m.cmp_contextual_help_related_docs()}</span>
-          </div>
-          <div className="docs-card">
-            {relatedDocs.map((doc, i) => (
-              <a
-                key={i}
-                className="docs-item"
-                href={doc.url}
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <div className="badge">
-                  <Icon
-                    icon={IconKind.ActivityNotes}
-                    size={20}
-                    staticColor={ThemeVariable.FgAction}
-                  />
-                </div>
-                <div className="link">
-                  <span>{doc.title}</span>
-                  <Icon
-                    icon={IconKind.OpenInNewWindow}
-                    staticColor={ThemeVariable.FgAction}
-                  />
-                </div>
-              </a>
-            ))}
-          </div>
-        </div>
-      )}
-
-      {bestPractices && (
-        <div className="contextual-help-section">
-          <div className="header">
-            <Icon icon={IconKind.LightBulb} />
-            <span className="title">{m.cmp_contextual_help_best_practices()}</span>
-          </div>
-          <div className="best-practices-card">
-            <RenderMarkdown content={bestPractices} />
-          </div>
-        </div>
-      )}
+      {isPresent(faqs) && <ContextualHelpFaqs faqs={faqs} />}
+      {isPresent(relatedDocs) && <ContextualHelpDocs docs={relatedDocs} />}
+      {bestPractices && <ContextualHelpBestPractices content={bestPractices} />}
     </div>
   );
 };
